refactor(codegen): extract constraints into plugin folder
