+++ scripts/codegen/src/core/registry.ts
@@ -1,5 +1,6 @@
 import { a11yPlugin } from "../plugins/a11y/index.ts";
 import { branchesPlugin } from "../plugins/branches/index.ts";
+import { constraintsPlugin } from "../plugins/constraints/index.ts";
 import { coveragePlugin } from "../plugins/coverage/index.ts";
 import { defaultChildrenPlugin } from "../plugins/defaultChildren/index.ts";
 import { dependenciesPlugin } from "../plugins/dependencies/index.ts";
@@ -46,4 +47,5 @@ export const PLUGINS: readonly SubstratePlugin[] = [
   rootElementStaticPlugin,
   eventsPlugin,
   vocabularyPlugin,
+  constraintsPlugin,
 ];
