+++ scripts/codegen/src/semantic-checks.test.ts
@@ -2,8 +2,6 @@ import { describe, expect, test } from "vitest";
 import type { Vocabulary } from "./lib/vocabulary.ts";
 import { Spec } from "./schema.ts";
 import {
-  checkConstraintsAgainstCoverage,
-  checkConstraintsAgainstExamples,
   checkRepeatingParts,
   checkStateMachines,
   levenshtein,
@@ -65,26 +63,6 @@ const vocabulary: Vocabulary = {
   },
 };
 
-function makeButton(overrides: Partial<Spec> = {}): Spec {
-  return Spec.parse({
-    name: "button",
-    kind: "atomic",
-    element: "button",
-    rootClass: "t-button",
-    variants: { solid: { description: "Filled." }, outline: { description: "Outlined." } },
-    intents: {
-      primary: { description: "Primary." },
-      danger: { description: "Danger." },
-    },
-    sizes: { sm: { description: "Small." }, md: { description: "Medium." } },
-    tokens: {
-      bg: { fallback: "--t-accent", desc: "Background." },
-      fg: { fallback: "--t-on-accent", desc: "Foreground." },
-    },
-    ...overrides,
-  });
-}
-
 /** Validates the literal at runtime via Zod. The lint rule
  *  `no-as-unknown-cast` forbids the bare schema-cast in test files (it
  *  hides drift); route every fixture through this helper so a renamed
@@ -106,71 +84,6 @@ describe("levenshtein + suggest", () => {
 
   test("suggest returns undefined when nothing is within range", () => {
     expect(suggest("destructive", ["danger", "success"], 3)).toBeUndefined();
-  });
-});
-
-describe("checkConstraintsAgainstExamples", () => {
-  test("flags an example that matches `when:` and uses a `forbid:` value", () => {
-    const spec = makeButton({
-      constraints: [
-        {
-          when: { variant: "outline" },
-          forbid: { intent: ["danger"] },
-          reason: "Outline-danger has no surface to apply to.",
-        },
-      ],
-      examples: [{ id: "outline-danger", props: { variant: "outline", intent: "danger" } }],
-    });
-    const issues = checkConstraintsAgainstExamples(spec);
-    expect(issues).toHaveLength(1);
-    expect(issues[0]?.path).toBe("examples.outline-danger");
-  });
-
-  test("passes a constraint whose `when:` does not match", () => {
-    const spec = makeButton({
-      constraints: [
-        {
-          when: { variant: "outline" },
-          forbid: { intent: ["danger"] },
-          reason: "n/a",
-        },
-      ],
-      examples: [{ id: "solid-danger", props: { variant: "solid", intent: "danger" } }],
-    });
-    expect(checkConstraintsAgainstExamples(spec)).toEqual([]);
-  });
-});
-
-describe("checkConstraintsAgainstCoverage", () => {
-  test("prunes constraint-violating cells (coverage expansion drops them)", () => {
-    const spec = makeButton({
-      constraints: [
-        {
-          when: { variant: "outline" },
-          forbid: { intent: ["danger"] },
-          reason: "No surface.",
-        },
-      ],
-      coverage: { variant: true, intent: true },
-    });
-    // Constraints prune the cell set before expansion. The check walks the
-    // pruned set; the violating cell (outline × danger) is excluded, so the
-    // check is silent.
-    expect(checkConstraintsAgainstCoverage(spec)).toEqual([]);
-  });
-
-  test("does not flag a coverage block that constraints leave alone", () => {
-    const spec = makeButton({
-      constraints: [
-        {
-          when: { variant: "outline" },
-          forbid: { intent: ["danger"] },
-          reason: "n/a",
-        },
-      ],
-      coverage: { variant: ["solid"], intent: true },
-    });
-    expect(checkConstraintsAgainstCoverage(spec)).toEqual([]);
   });
 });
