+++ scripts/codegen/src/semantic-checks.ts
@@ -8,6 +8,10 @@ import type { TokenDictionary } from "./lib/token-dictionary.ts";
 import type { Vocabulary } from "./lib/vocabulary.ts";
 import { checkA11yRefs } from "./plugins/a11y/check.ts";
 import { checkBranches } from "./plugins/branches/check.ts";
+import {
+  checkConstraintsAgainstCoverage,
+  checkConstraintsAgainstExamples,
+} from "./plugins/constraints/check.ts";
 import { checkCoverageShape } from "./plugins/coverage/check.ts";
 import { checkElementByProp } from "./plugins/elementByProp/check.ts";
 import { checkEvents, checkEventsRuntimeSupport } from "./plugins/events/check.ts";
@@ -45,12 +49,7 @@ type CssIndex = {
   cssByName: Map<string, string>;
 };
 
-type AtomicSpec = Spec & { kind: "atomic" };
 type CompositeSpec = Spec & { kind: "composite" };
-
-function isAtomic(spec: Spec): spec is AtomicSpec {
-  return spec.kind === "atomic";
-}
 
 function isComposite(spec: Spec): spec is CompositeSpec {
   return spec.kind === "composite";
@@ -101,144 +100,6 @@ export function suggest(
 function suggestionFragment(name: string, options: readonly string[]): string {
   const hint = suggest(name, options);
   return hint ? ` Did you mean '${hint}'?` : "";
-}
-
-// ── Constraint enforcement (examples + coverage cells) ──────────────────────
-
-type Constraint = NonNullable<AtomicSpec["constraints"]>[number];
-
-/**
- * Apply a constraint to a candidate cell. Returns the matched forbid key/value
- * pair when the cell violates the constraint, or undefined when the cell is
- * clean.
- */
-function violation(
-  cell: Record<string, unknown>,
-  constraint: Constraint,
-): { prop: string; value: unknown } | undefined {
-  for (const [prop, expected] of Object.entries(constraint.when)) {
-    if (cell[prop] !== expected) return undefined;
-  }
-  for (const [prop, forbidValue] of Object.entries(constraint.forbid)) {
-    const cellValue = cell[prop];
-    if (cellValue === undefined) continue;
-    const forbiddenList = Array.isArray(forbidValue) ? forbidValue : [forbidValue];
-    if (forbiddenList.some((v) => v === cellValue)) return { prop, value: cellValue };
-  }
-  return undefined;
-}
-
-function describeConstraint(constraint: Constraint): string {
-  const whenStr = Object.entries(constraint.when)
-    .map(([k, v]) => `${k}=${JSON.stringify(v)}`)
-    .join(", ");
-  return `constraint (when ${whenStr}): ${constraint.reason}`;
-}
-
-export function checkConstraintsAgainstExamples(spec: Spec): Issue[] {
-  const issues: Issue[] = [];
-  if (!isAtomic(spec)) return issues;
-  const constraints = spec.constraints ?? [];
-  if (constraints.length === 0) return issues;
-  for (const example of spec.examples ?? []) {
-    const cell = example.props ?? {};
-    for (const constraint of constraints) {
-      const v = violation(cell, constraint);
-      if (v !== undefined) {
-        issues.push(
-          issue(
-            spec.name,
-            `examples.${example.id}`,
-            `violates ${describeConstraint(constraint)}; '${v.prop}=${JSON.stringify(v.value)}' is forbidden`,
-          ),
-        );
-      }
-    }
-  }
-  return issues;
-}
-
-/**
- * Cartesian-expand `coverage:` into cells, pulling each dimension's values
- * from the spec's `variants:` / `intents:` / `sizes:` / `props:` / `states:`
- * declarations. Dimensions declared as a string-list (e.g.
- * `states: [disabled, loading]`) use that subset; `true` means "every
- * declared value". The codegen path uses pairwise reduction for fixtures; a
- * cartesian walk is fine here because we only need it for constraint checks.
- */
-function expandCoverage(spec: AtomicSpec): Record<string, unknown>[] {
-  const coverage = spec.coverage;
-  if (!coverage) return [];
-  const dimensions: { name: string; values: unknown[] }[] = [];
-  for (const [dimName, declaration] of Object.entries(coverage)) {
-    const declared = collectDimensionValues(spec, dimName);
-    const values = declaration === true ? declared : Array.isArray(declaration) ? declaration : [];
-    if (values.length === 0) continue;
-    dimensions.push({ name: dimName, values });
-  }
-  if (dimensions.length === 0) return [];
-  let cells: Record<string, unknown>[] = [{}];
-  for (const dim of dimensions) {
-    const next: Record<string, unknown>[] = [];
-    for (const cell of cells) {
-      for (const value of dim.values) next.push({ ...cell, [dim.name]: value });
-    }
-    cells = next;
-  }
-  return cells;
-}
-
-function collectDimensionValues(spec: AtomicSpec, dim: string): string[] {
-  switch (dim) {
-    case "variant":
-      return Object.keys(spec.variants ?? {});
-    case "intent":
-      return Object.keys(spec.intents ?? {});
-    case "size":
-      return Object.keys(spec.sizes ?? {});
-    case "visualStates":
-      return Object.keys(spec.visualStates ?? {});
-    default: {
-      const propDef = spec.props?.[dim];
-      if (propDef?.values) return propDef.values;
-      return [];
-    }
-  }
-}
-
-/**
- * Asserts the coverage expansion is internally consistent with constraints.
- * Constraints prune the cell set before expansion — forbidden cells never
- * enter the candidate set. This check walks the post-pruning cells and
- * confirms none violate a constraint: a defensive sanity gate that catches
- * a future regression in pruning logic.
- */
-export function checkConstraintsAgainstCoverage(spec: Spec): Issue[] {
-  const issues: Issue[] = [];
-  if (!isAtomic(spec)) return issues;
-  const constraints = spec.constraints ?? [];
-  if (constraints.length === 0) return issues;
-  const cells = expandCoverage(spec).filter(
-    (cell) => !constraints.some((c) => violation(cell, c) !== undefined),
-  );
-  for (const cell of cells) {
-    for (const constraint of constraints) {
-      const v = violation(cell, constraint);
-      if (v !== undefined) {
-        const cellStr = Object.entries(cell)
-          .map(([k, val]) => `${k}=${JSON.stringify(val)}`)
-          .join(", ");
-        issues.push(
-          issue(
-            spec.name,
-            "coverage",
-            `expanded cell {${cellStr}} violates ${describeConstraint(constraint)}; '${v.prop}=${JSON.stringify(v.value)}' is forbidden`,
-          ),
-        );
-      }
-    }
-  }
-  return issues;
 }
 
 export { checkA11yRefs };
@@ -1134,6 +995,10 @@ export function runSemanticChecks(
 }
 
 export { checkBranches } from "./plugins/branches/check.ts";
+export {
+  checkConstraintsAgainstCoverage,
+  checkConstraintsAgainstExamples,
+} from "./plugins/constraints/check.ts";
 export { checkCoverageShape } from "./plugins/coverage/check.ts";
 export type { DependencyIndex } from "./plugins/dependencies/check.ts";
 export { checkDependencyCycles } from "./plugins/dependencies/check.ts";
